+++ core/src/fr/limsi/rorqual/core/dpe/enums/generalproperties/TypeEquipementEclairageEnum.java
@@ -4,16 +4,16 @@ package fr.limsi.rorqual.core.dpe.enums.generalproperties;
  * Created by ricordeau on 30/07/15.
  */
 public enum TypeEquipementEclairageEnum {
-    UNIQUEMENT_BASSE_CONSOMMATION(1.5),
+    UNIQUEMENT_BASSE_CONSOMMATION(1.5f),
     MAJORITAIREMENT_BASSE_CONSOMMATION(2),
-    MAJORITAIREMENT_AMPOULE_A_INCANDESCENCE(3.7);
+    MAJORITAIREMENT_AMPOULE_A_INCANDESCENCE(3.7f);
 
-    private double consommationEclairage;
-    TypeEquipementEclairageEnum(double conso){
+    private float consommationEclairage;
+    TypeEquipementEclairageEnum(float conso){
         this.consommationEclairage=conso;
     }
 
-    public double getConsommationEclairage(){
+    public float getConsommationEclairage(){
         return this.consommationEclairage;
     }
 }
